[honey] Index valuestream chunks by the last docid
Then we can use find_entry_ge() in skip_to() which means jumping into a
gap between chunks puts us on the chunk after the gap.  With the
previous approach, we'd end up on the chunk before.

The equivalent change has already been made for posting chunks and
doclen chunks.