Index POD files despite minor formatting errors
Pass --errors=stderr to pod2text so that minor formatting errors don't
prevent us from indexing a file.

It may seem that --errors=none is a better option, but at least with
current versions of Pod::Text that results in an ERRATA section in the
generated text version, which we then end up indexing.