[honey] Merge cursor find_exact() and find_entry_ge()
They're currently basically the same, but pass a flag to say which
we are doing, as there's scope for doing less work for find_exact()
in some cases for certain index structures (e.g. with an array index
we can determine that a key isn't present if there are no keys which
start with that particular first character).