Fix glass freelist bug
If changes to a new database which didn't modify the termlist table
were committed, then a block which had been allocated to be the root
block in the termlist table was leaked.  This was largely harmless,
except that it was detected by Database::check() and caused an error.

Reported by Antoine Beaupré and David Bremner.