Range check winsock2's socket() return value
Winsock2's socket() returns the unsigned type SOCKET, which is a 32-bit
type for WIN32 and a 64-bit type for WIN64.

It seems we can always safely assign SOCKET to an int: failure is indicated
by INVALID_SOCKET which will cast to -1 as an int, and it seems in
practice that valid values all fit in 31-bits (and that we're not the
only code to assume this since it makes it much easier to write code
that deals with BSD sockets and winsock2's bastardised version of them)
so Microsoft are unlikely to arbitrarily change that).

But we should check and throw an exception rather than quietly mangling
the value.

See #738.