[honey] For bchop index store upper bound key
Add a key which is an upper bound on all keys in the table, pointing
to the root (start of the index) so that searching for a key larger
than any in the table quickly fails.

If the largest key in the table has 0xff for all its first four bytes
then we don't add this new entry (because there's no 4 byte entry
which gives an upper bound).  It's probably less likely that keys
beyond the end of the table will be searched for in this case anyway.