[honey] Don't allow mixed wdf for a term
A given term in a honey database must either always have wdf=0 (i.e.
a boolean term) or never have wdf=0.

Currently we make use of this to reduce the size of the postlist header
and to avoid storing wdf explicitly when we can deduce it must be 1 for
the second and subsequent occurrence.