bootstrap: Preserve options which should be sticky
We may automatically rerun bootstrap to regenerate the top-level
configure.  When we do, we now preserve options used when bootstrap
was originally run (except for --clean which doesn't seem like it
should be sticky, and --without-autotools which only warns that it
is obsolete).

(cherry picked from commit 74253269450ef16827d302740d0202e8812d95f8)