Account for fd inside DIR in fd leak tracking
We open /dev/fd or /proc/self/fd using opendir to find out what fds are
open, but the DIR that returns wraps an fd open on the directory which
means we see an extra fd which can confuse the tracking.  Use dirfd() to
find this fd and ignore it.  When dirfd() isn't available just disable
fd tracking (it's specified by POSIX, so it ought to be widely
supported).