Make TermIterator::get_termfreq() always exact
Previously for multi-databases we approximated the result, which is
probably either a hang-over from when this method was used during
Enquire::get_eset(), or else due to a thinking that this method would be
used in that situation (it certainly is not now).  If the user creates
a TermIterator object and asks it for term frequencies then we really
should give them the correct answer - it isn't hugely costly and the
documentation doesn't say it might be approximated.