[honey] Reform doclen and boolean term chunks
When converting a glass DB to honey, reform doclen chunks (which are
about 30% smaller in honey) and merge pairs of boolean term chunks
(which are very close to 50% smaller in honey).

This doesn't have much of an impact on postlist table size (because
we prefix-compress keys) but it should speed up scanning through a
table.