Allow resolve_duplicate_metadata() to delete entry
Previously returning an empty string in an implementation of
Compact::resolve_duplicate_metadata() would result in a user meta-data
entry with an empty value, which isn't normally achievable (empty values
aren't stored), and so will cause odd behaviour.  We now handle an empty
returned value by interpreting it in the natural way - it means that the
merged result is to not set a value for that key in the output database.

(cherry picked from commit ea1ef7cc65c29cb0487419c3096f28708d19f8ca)