Conditionalise move semantics support
Define XAPIAN_MOVE_SEMANTICS if we're confident the compiler in
use supports them.

C++11 move semantics are very useful in threaded code that wants to
hand-off Xapian objects to worker threads, but in this case it's very
unhelpful for availability of these semantics to vary by compiler as it
quietly leads to a build with non-threadsafe behaviour.

To address this, user code can #define XAPIAN_MOVE_SEMANTICS to force
this on, and will then get a compilation failure if the compiler lacks
suitable support.