Use a brief description for xapian-core.pc.
According to the pkg-config(1) man page the 'Description' field
should be brief and pkgconf will print warnings with --validate. A
multi line 'Description' is not valid and neither pkg-config or
pkgconf will print more than the first line with --list-all.

This description is better suited for a man page or README.

See:

  Description:
    This should be a brief description of the package

  https://linux.die.net/man/1/pkg-config