Improve API docs for MSet::snippet()
Clarify why non-query terms are considered.

Explain more clearly that slightly more bytes might be selected than the
length parameter specifies, and that this is before escaping,
highlighting, and omission markers.

Note the escaping bug in 1.4.5 and earlier.

Clarify that "[t]he stemmer used to build the query" means "same
stemming algorithm" - it doesn't need to be the exact same Stemmer
object.