Call stat() after mkdir() to avoid race
If we stat the path first to see if the directory already exists
then with multiple callers trying to perform the same operation we
can get one creating the directory between another calling stat()
and mkdir(), which would lead to the second caller failing even
though the directory actually exists.

We're seeing occasional spurious failures under Cygwin here, and
I'm wondering if this is a factor.