Fix configure probes for builtins with clang
With clang we need to specify the argument types for each builtin,
or else AC_CHECK_DECLS tries to compile code which just references
the value of the builtin function causing clang to give an error
saying that's not allowed.  If the argument types are specified
then AC_CHECK_DECLS tries to compile a call to the builtin function
instead.