Tweak split positions code to avoid any size bloat
Make the split point an unsigned int instead of a size_t so the
structure is the same size as before this change.  It's possible
with 64-bit termpos for the split point to not be representable,
but only when there are more than 4 billion postings for the same
term in a single document, which isn't realistically going to
happen in non-pathological data.  If we can't store the split
point, we fall back to inserting into the vector like we did
before the split positions optimisation.

(cherry picked from commit 585a17a0117c7d3e370ce49f4c1efbb16eddb596)