Add new stemming mode STEM_SOME_FULL_POS
This stores positional information for both stemmed and unstemmed
terms, and allows NEAR and ADJ to work with stemmed terms.

The extra positional information is likely to take up a significant
amount of extra disk space so the default STEM_SOME is likely to be
a better choice for most users.