Adjust --mime-type to allow ':' in extension
A valid MIME Content-Type can't contain a colon, so if the argument
to --mime-type contains more than one colon it makes more sense
to split at the *last* colon (we used to split at the first), as
an extension could conceivably contain a colon.

Mostly this change is for consistency with the new --mime-type-match
option, where the leafname pattern could reasonably contain a colon.

(cherry picked from commit 67aaaa6507b65787341973efc7db880812afcbc4)