Fix OP_SYNONYM freq ests over OP_XOR or OP_MAX
When OP_SYNONYM was used with a subquery containing OP_XOR or OP_MAX
the termfreq, reltermfreq and collfreq were incorrectly estimated
in several ways.

The frequency estimates for the first subquery of the OP_XOR/OP_MAX
was being used for all subqueries.

And much more minor, the estimated collfreq was always being rounded
down rather than rounded to the nearest integer.