[scriptindex] Drop fallback if replace_document fails
Ever since the initial version of scriptindex.cc added back in 2002
by 782e9d06729866aefa3fa8a1161d0ac4d6db5981 there's been code to
catch Xapian:Error from replace_document() and fall back to using
add_document() with the comment:

Possibly the document was deleted by another process in the meantime...?

This comment doesn't make any sense as we only support a single writer
at a time.  Digging back further in the history, there's similar code in
the old "indexnasty" example which scriptindex was a replacement for,
so this probably just comes from a time before we'd entirely settled
on the single writer model.

The only exception which modern WritableDatabase::replace_document()
can throw which it would be useful to do this for is
Xapian::InvalidArgumentError if a zero docid is specified, but we
explicitly exclude that case before calling replace_document().