Try to flag file as failed on std::bad_alloc
If we catch std::bad_alloc while trying to index a file that's a
sign that it may be a pathological case.  Rather than letting the
indexing get stuck on such files it makes more sense to flag them
as failed so they are skipped on future runs like other problematic
files (assuming we don't run out of memory again trying to commit
the changes).