[PHP7] Fix to build with PHP 7.3
In PHP 7.3 array_init() changed from a function returning int (in fact
always the constant SUCCESS) to a macro expanding to a do/while loop.

We were checking the return value, so our code failed to compile with
PHP 7.3.  Just drop the check - it wasn't useful with older PHP either
since the return value was always SUCCESS.