Support latest Travis macOS build
Homebrew has for a while shipped python as the python3 package (with
python2 for py2), but Travis was a little behind for a while. It
seems to have caught up, which means that upgrading the python
package before installing python2 is no longer needed.

(cherry picked from commit bcb093ba7f73d59337943e5064fdb1af4526ebbe)