Fix bug in split position handling
When we look to see in the new position is present before the split
we were only checking if the std::lower_bound() (binary chop) ended
up on a valid position, but we also need to check if that position
is the one we're trying to add.

This was leading to positions getting lost in this situation.

(cherry picked from commit edc381036725373818fa04d137a69f5941a95150)