Use poll() instead of select() if available
poll() is specified by POSIX.1-2001 so should be widely available
by now, and it allows watching any fd (select() is limited to watching
fds < FD_SETSIZE).

For any platforms which still lack poll() we now workaround this
select() limitation when a high numbered fd needs to be watched (for
example, by trying a non-blocking read or write and on EAGAIN sleeping
for a bit before retrying).