Use libtoolize path to determine aclocal dir
Previously we used the path to the libtool script, but in the
Debian packaging that's in libtool-bin which may not be installed
(and our build works without it).  We definitely require libtoolize
to be present, so use that instead.

We also now fail with a clear error if libtoolize isn't found on
PATH - previously we'd die with an obscure error due to "aclocal
-I -I /foo" being run, which results in the first -I taking the
second as its argument, and then "/foo" being interpreted as a
command line argument.

(cherry picked from commit cd25e7be5cb0c45ebd5675de92d4e742ef0786c8)