Add tests for BAD_VALUENO wrapping
This ought to be wrapped as a constant everywhere, but it only is
for Python.  For some languages, changing to a constant won't change
how user code looks, but where that does change we want to keep existing
user code working, so start by adding feature tests for the existing
wrapping.