backport-source-packages: Fix for dpkg 1.19.3
With dpkg 1.19.3 we get a *_source.changes which doesn't have a
"Binary:" field, which is then rejected when one tries to upload
it to launchpad.

We work around this by adding "Binary: dummy", which gets us
past the upload check, and it doesn't actually matter that this
doesn't match the binary packages which are produced by the build.