Don't always open logfile with O_SYNC
We were doing this so that the most recent log entries are less
likely to be lost if there's a crash, but O_SYNC can incur a significant
performance overhead and most debugging is not of such crashes.

You can now request synchronous logging by including %! in the filename
specified with XAPIAN_DEBUG_LOG (which is replaced with the empty
string).

We also now use O_DSYNC if available in preference to O_SYNC, since
the mtime of the log file isn't important.

(cherry picked from commit c5e74a93974f848fe9defd0296243eed222a2147)