Improve OP_ELITE_SET handling
OP_ELITE_SET actually recursively explodes any OP_OR (or OP_OR-like)
subqueries into the list of subqueries it selects from, which is
not supposed to happen.  But until that's fixed, we now select from the
full exploded list rather than the last n (where n is the number of
direct subqueries of the OP_ELITE_SET).