Add SNIPPET_CJK_NGRAM flag
This adds working support for selecting and highlighting snippets
which works with the QueryParser and TermGenerate FLAG_CJK_NGRAM
flags.  This mode can also be enabled by setting environment
variable XAPIAN_CJK_NGRAM to a non-empty value.

There was nominally already support for XAPIAN_CJK_NGRAM in
MSet::snippet(), but it didn't work usefully - the highlighting
added was all empty start/end pairs at the end of the span of CJK
characters containing the CJK ngram terms.  To the user it would
typically look like it was selecting the end of the text and not
highlighting anything.