Update to use SWIG pre-4.0.0
SWIG is approaching 4.0.0 so update git master to use the current
state of development of this so we can help shake out any issues.

This should fix issues with recent PHP 7.x releases.

This requires dropping support for Python 2.6, but it was time to
do that anyway.

(cherry picked from commit 483b3a2f50bb26e74821967585b8bf0df7368ad5)