bootstrap: Unshallow the tree if necessary
In a shallow clone we may not have the commit needed to set up common
for omega and letor.  We now check if the commit is present, and if
not and the clone is shallow then run `git fetch --unshallow`.

This means we can stop setting XAPIAN_COMMON_CLONE_URL for CI, and
fix PRs which change common and need to use those changes in omega or
letor (problem discovered by Vaibhav Kansagara).