Handle an empty value slot lower bound gracefully
This shouldn't happen for a non-empty slot, but has been reported
by a notmuch user so it seems there is (or perhaps was as the database
was several years old) a way it can come about.  We now check for
this situation and set the smallest possible valid lower bound
instead, so other code assuming a valid lower bound will work
correctly.  Reported by jb55.