omegatest: Tweak to not fail when $time not supported
If the OS time functions report an error, $time returns an empty
value.  This seems to happen on GNU Hurd for some reason, and
causes a testcase to produce different output to that expected.
Adjust the testcase so that this event doesn't cause it to fail.

(cherry picked from commit 6c332d5b48b17b1f54608f9798e89d73dd03b364)