Support mixed Chinese numbers within CJK texts
it is a partial patch for the ticket #699 mixed CJK number,
it is implemented in the operator++ of CJKNgramIterator,
and works only for mixed number start with Latin digit
and the previous character should be a CJK character.
Reasons for these limits are puzzles to the requirements
and considerations on performence. So maybe it would change quickly.