Optimise edit distance calculation more
We're always calculating edit distances to a fixed target, so we
can lazily allocate the amount of space we'll need, and pre-compute
the frequency histogram for the target.

The length-based check before we convert the candidate to UTF-32
now uses the UTF-32 length of the target, which makes it a bit
more effective.