Suppress warnings from older Perl and Ruby headers
The warnings are due to use of register and constructs which look
like C++11 user-defined literals.  They're fixed in newer versions
so they're just noise in our build.

We were working around these in the CI build, so drop that workaround
as we want to make the build warning-clean for users too.

Reported by daniel93 on #xapian.