Allow filter commands to take input on stdin
Most of the external filters we have built-in support for can, and this
avoids having to open the input file an extra time (omindex needs to
open the input file so it can calculate a checksum of the contents for
duplicate detection, and also may need to use libmagic to find the
file's MIME Content-Type).

In the future it will probably also allow extracting text from documents
attached to emails, in ZIP files, etc without having to write them to a
temporary file just to have a file to run the filter command on.