Trim HACKING and update README
A lot of the contents of xapian-core/HACKING has been moved into
the Xapian developer guide, so remove that (to avoid the two copies
getting out of sync) and add a pointer to the guide.

Mention both developer and user guide in README, since they're
probably the most useful places to get started now.