Make Document updating a bit more efficient
Use std::move to avoid copying of OmDocumentTerm objects.  This
improvement is not relevant for git master as this code has been
replaced.

Use std::map::emplace_hint() with a hint of end() when bulking inserting
in sorted order so that the bulk insertion is O(n) instead of
O(n*log(n)).  The equivalent change was made to git master in
611754f0c848e3ed5a5dc77ed78cc56e0cbf869e.