Fix socket() wrapper
We add a wrapper around the libc socket() function which deals with
the corner case where SOCK_CLOEXEC is defined but socket() fails if
it is specified (which can happen with a newer libc and older kernel).

Unfortunately, this wrapper wasn't checking the returned value from
socket() correctly, so when SOCK_CLOEXEC was specified and non-zero
it would create the socket() with SOCK_CLOEXEC, then leak that one
and create it again without SOCK_CLOEXEC.

Now we handle the return value check properly.