Adjust for timezone in parsedate %z format
The timezone information comes through in the non-standard tm_gmtoff
member of struct tm, which seems to mean we can't support this on
platforms without this, so flag an error when %z is used on such
platforms.

Also flag an error for %Z everywhere as that doesn't seem to be
usefully supported.