[python3] Raise TypeError for bad Unicode input
Previously cases such as a lone surrogate would be handled by quietly
skipping the bad codepoints when converting to UTF-8 to pass to Xapian.

A bonus effect of this fix is we no longer use the deprecated old-style
Py_UNICODE API, which currently gives deprecation warnings and is
slated to be removed in Python 4.0.