[perl] Implement manual refcounting for add_(boolean_)prefix
The perl bindings resort to manual reference counting to ensure that
temporary objects that are passed to the C++ side for use in callbacks
do not get destroyed too early. To this end, a proxy method intercepts
the calls and registers the references with the potential callback
caller.

This commits enables said behaviour for registering FieldProcessor
instances within add_prefix and add_boolean_prefix.