[perl] Enable passing subs as FieldProcessors
To use a custom FieldProcessor, the user has to generate a new
PerlPackage as a subclass for every new processor. This process can be
simplified by letting the caller use subroutinges as arguments to
Xapian::QueryParser::add_prefix and
Xapian::QueryParser::add_boolean_prefix that are automatically wrapped
to Classes by the api.