scriptindex: Emit warning for invalid date action value
The documentation used to say "invalid values are ignored at present"
which was accurate before the recent integer parsing changes made this
case an error which stopped indexing.  It's useful to report this, but
more in keeping with behaviour elsewhere to make this a warning so that
a single bad value doesn't block indexing of otherwise good data.

(cherry picked from commit d7b8907228dd1db7624173f6e722349550ffe5a0)