Add testcases for feature related api's.
These include unit testcases for the 5 features such as TfFeature,IdfFeature,
TfIdfDoclenFeature,CollTfCollLenFeature,TfDoclenCollTfCollLenFeature.
Additionally it includes test for FeatureList class which includes testing
an empty featurelist and a featurelist containing more than it's default
features and also for generating a database with three documents in which
one has no common terms with other two.