Fix merging empty MSet objects
If there are both local and remote databases, or more than one
remote database, then we merge the resulting MSet objects.  If
they were all empty then the merge would try to call front() on
an empty vector.

No regression test here, but Tanmay's working on test harness
changes which will provide a regression test (it was these changes
which found the bug).