Change feature API design
This includes adding PIMPL approach for feature class and changing how
the stats needed for featurelist are populated and computed. After this
design change all the features in the featurelist share a common
Feature::internal object which greatly decreases the memory required as
maps related to statistics do not need to be copied for every feature.