Ensure output directory in generate_exceptions
For -letor, this means we don't have to create the output
directory in configure. For -core we still do, because it gets
used within the configure script itself (to make version.h). To
keep generate_exceptions scripts reasonably in sync, the -core
version will ensure the output directory exists even though it
always should do.