Add Database::size() method
This method reports the number of shards.

This change also improves behaviour in the corner case of adding
a multi-db to an empty database.  Previously this had the same
effect as assigning the database object would.  Now the list of shards
is copied instead, which seems more consistent with what happens when
adding to a non-empty database.