Fix score method
Add only those FeatureVector which are found in the qrel file. Earlier
the docid's in the mset which were not present in the qrel file also
contributed to featurevectors whose label values were used to determine
the NDCG score values and their labels remained undefined and as a result
gave unexpected values for NDCG scores like nan.