Adjust scriptindex deletion handling rules
Previously scriptindex would delete a document if it only had input
fields which only match a line in the index script containing the
"unique" action.

This is problematic if you want to process the same input field in
two different ways using commands which modify the current text -
for example, `hash unique=Q boolean=Q` and `split=/ boolean=XP` can't
be combined into a single line of actions with the same effect as
each on a separate line.

So now the check is based on whether there are any input fields
which don't match a line in the index script containing the "unique"
action, and this case can be handled by writing it as two lines.