Fix honey test harness bug
Currently generated honey databases are built as glass databases and
then converted by compacting.  However the compaction wasn't being done
with Xapian::DBCOMPACT_NO_RENUMBER so if there was a gap at the start
of the used docids then they were renumbered, which was breaking
testcase compactnorenumber1.

The equivalent fix is also applied for single-file glass databases,
but we don't yet support generated databases there so that wasn't
being triggered.