Update documentation about how to add a new format
Currently, there is an outdated FAQ entry where it is explained how to
add a new file format to omindex. We can update this information and
explain how to add this kind of support to Omega through a library or an
external filter.

(cherry picked from commit 6543aa69e84b5581f2b2c929582bdbc85ac59506)