Change how PostingSource tells matcher maxweight changed
Previously each PostingSource object stored a pointer to the internal
PostListTree object in a void* pointer, and internally this was cast to
PostListTree* and a particular method called on it, which cleared a bool
member in the PostListTree object.

The new approach is to store a pointer to that bool in each
PostingSource instead.  This seems cleaner, and also allows
set_maxweight() to be declared inline in the public API header.