Improve lower bound when collapsing and 0 hits requested
We were returning a lower bound of zero in this case, but actually
we know there must be at least the minimum of the lower bound from
the stats and collapse_max (the worst case is that all the hits
have the same collapse key).

Bug found and identified by Tanmay Sachan with help from Guruprasad
Hegde.  Tanmay's upcoming testsuite changes will provide a regression
test for this bug.