Fix assumption in sort_existing_mset_by_relevance
This testcase was assuming the max_possible value in the MSet, but
that depends on the wdf upper bound reported by the backend.  If
the test is run with a sharded remote database, the remote backend
currently reports a less tight wdf upper bound than local backends,
and this testcase fails.

Problem found by Tanmay Sachan while working on test harness
improvements.