Fix Ranker API
Before we were dealing with a vector of FeatureVector objects, ie one
FeatureVector per entry in the training set and queryids were completely
ignored. Now we have a separate vector per query in the training set.

For more info on why this change is needed for the implementation of
ListNetRanker look at page 5 of:
https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/tr-2007-40.pdf

For the implementation of ListMleRanker look at page 6 of:
http://icml2008.cs.helsinki.fi/papers/167.pdf

If you look closely you will notice that the current implementation
doesn't take query into account which is clearly wrong.