[ci] Enable testing with PHP5 and PHP7
We were disabling PHP7 for the GCC build because we force a build
with GCC 4.7 (as that's the baseline GCC version we aim for support
for 1.4.x) but PHP7 fails to work with that because it assumes
builtins are available based on tests done when PHP7 was built.

To address this, add a new build with the standard GCC version,
and install php5 packages for this build so it tests PHP5 too
(PHP5 is dead upstream, but given we're continuing to ship our
PHP5 bindings for 1.4.x it's useful to know they still build).