Avoid warning from GCC 9
GCC 9 with -Wdeprecated-copy (which is enabled by -Wextra) gives a
warning if a reference to an Error object is thrown because the
default assignment operator is suppressed but an implicit default
copy constructor is needed, and that situation was deprecated in
C++11.  We can avoid this by explicitly defaulting Error's copy
constructor.

(cherry picked from commit 311ed168df1231b054a34f561ab1bb32e9e1a0d1)