Initialize the parameters for neural network with Xavier initialization
This change applies to both ListNetRanker and ListMleRanker.
The motivation for Xavier initialization in Neural Networks is to
initialize the weights of the network so that the neuron activation
functions are not starting out in saturated or dead regions. In other
words, we want to initialize the parameters with random values that are
not “too small” and not “too large.”