Fix match stats when collapsing with remote shards
When searching with collapsing over multiple shards and at least some
are remote, then the bounds and estimate for the number of matching
documents weren't being handled correctly in some cases.

The fix for this requires us to request check_at_least matches from
each shard in this case.

Bugs uncovered by Tanmay Sachan's upcoming change to enable running
generated testcases for sharded databases with remote or mixed
remote+local shards, which will serve as a regression test for this.