Makefile.PL: fix automated testing false negative
In the case of Search::Xapian 1.2.x and xapian-core 1.4.y, $msg is
deliberately unset because the two versions should be
compatible. Since this is intended to be an escape from this
conditional, it should be respected by the AUTOMATED_TESTING check, so
automated testing with compatible libraries doesn't nevertheless
result in no Makefile.