Fix indexing of positions in CJK corner case
If indexing CJK ngrams, a stemmer is in use, and the stemming mode
isn't STEM_SOME then we'd add bigrams with positional information
(whereas only unigrams are meant to have positional information).

We don't provide any stemming algorithms for CJK languages, though
someone might enable this situation if indexing a mixture of CJK and
text in a language there's a stemmer for.  Or they might have a
custom stemmer.

The typo which caused this is also present in 1.4.x, but there it's
harmless because the CJK n-gram code is different.