Fix error handling in $lookup
We now check for an error from cdb_init() - previously we'd try to use
the struct cdb and then call cdb_close() on it, which would probably
crash.

We also now correctly check for an error from cdb_get().  Previously
completely the wrong variable was checked.  That other variable will
always be non-zero at this point, so there's no problem unless cdb_get()
fails.  This bug was introduced by a warning fix in 1.2.20.

Failures from cdb_init() and cdb_get() seem hard to trigger on demand,
so no regression test for these fixes.