Add API to reconstructing text from positions
This gives the text without punctuation or capitalisation, but it's
sometimes useful.  The key reason for putting this in the API is that
it allows us to handle the reconstruction on the server for remote
databases, which is much more efficient than opening a remote position
list for every term in the document.