Enhance Xapian::Stem API
The constructor which takes a stemmer name now takes an optional
second parameter - if this is true, then an unknown stemmer name
falls back to the "none" stemmer.  This allows simply constructing
a stemmer from an ISO language code without having to worry about
whether there's a stemmer for that language, and without having
to handle an exception if there isn't.

Add a new Stem::is_none() method which tests if this is a "none"
stemmer.