Fix failing multi_glass_remoteprog_glass tests on x86
When the tests are run under valgrind, remote servers should be run
using the runsrv wrapper script, but this wasn't happening for
remote servers in multi-databases - now it is.

Also, previously runsrv only used valgrind for the remote for an x86
build that didn't use SSE, but it seems there are x87 instructions in
libc that are affected by valgrind not providing excess precision, so do
this for x86 builds which use SSE too.

Together these changes fix failures of topercent2, xor2, tradweight1
under backend multi_glass_remoteprog_glass on x86.