Implement thread-safe move for intrusive_ptr_nonnull
To do this, this allow the internal pointer to be NULL in a moved-from
intrusive_ptr_nonnull.  Once moved from, the only valid operations are
to destroy the smart pointer, or to assign or move assign to it (after
which all operations are valid again).

Fixes #788, reported by Germán M. Bravo.