Simplify probes for snprintf
Stop worrying about the broken snprintf in libbsd in Linux libc4 -
that's from ~25 years ago so way too ancient to matter now.  We also no
longer use SNPRINTF_ISO anywhere so stop probing for it.  Callers of
SNPRINTF need to handle it responding to an undersize buffer by
returning either -1 (pre-ISO) or the required buffer size (ISO).

These changes simplify the configure probes and also mean we no longer
need to be able to run a test program to see how snprintf behaves, which
is more cross-compile friendly.