Fix segfault for corner case synonym query
If a synonym query has no matches and there are subqueries after it
then the query optimiser can attempt to use a hint postlist which
has already been deleted.  The reproducer involves a database with
no documents in.  I'm not sure that's the only case it can happen
in, but adding a single document with other terms in to the reproducer
stops the segfault.

This doesn't affect 1.4.x - I believe it was introduced in
643bfeb998b3abe3c51cf85c98b8a39fec13753a.