Avoid deprecation warning on recent Linux
We were including sys/sysctl.h if it existed, which it does on Linux
but we don't actually use it there.  Including it now warns that it
is deprecated, so skip including it under Linux.

Reported on IRC by kumaran.

(cherry picked from commit 5d086881299cd5a93f8382daa01ed533b9e70a29)