Don't swallow exceptions from close() on remote
This is especially unhelpful if we aren't in a transaction and so
try to commit() and that fails - previously this would look like
a successful close() to the caller.

No regression test as we don't currently have a mechanism to force an
error from a remote.

(cherry picked from commit 0236eab5077b20a458d0096689f6c36428c44349)