Fix phrase weight check optimisation
We need to make sure we fetch the stats for the correct docid, and
we should be checking the weight contribution from the sublist, not
for the whole query.

This could only happen for queries with a positional subquery whose
check couldn't be hoisted to the top level.  Generally it probably
only resulted in giving extra weight to some documents, but in
some situations when searching multiple shards it could result in
a bogus exception like "Document 143165576 not found" or an
assertion failing.